add minor improvements for entry tags
