+++ aries/aries_vcx_core/src/wallet2/entry_tag.rs
@@ -1,32 +1,9 @@
-use std::collections::HashMap;
-
 use serde::{Deserialize, Serialize};
 
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq, PartialOrd, Ord)]
 pub enum EntryTag {
     Encrypted(String, String),
     Plaintext(String, String),
-}
-
-#[cfg(feature = "vdrtools_wallet")]
-impl From<EntryTag> for (String, String) {
-    fn from(value: EntryTag) -> Self {
-        match value {
-            EntryTag::Encrypted(key, val) => (key, val),
-            EntryTag::Plaintext(key, val) => (format!("~{}", key), val),
-        }
-    }
-}
-
-#[cfg(feature = "vdrtools_wallet")]
-impl From<(String, String)> for EntryTag {
-    fn from(value: (String, String)) -> Self {
-        if value.0.starts_with('~') {
-            EntryTag::Plaintext(value.0.trim_start_matches('~').into(), value.1)
-        } else {
-            EntryTag::Encrypted(value.0, value.1)
-        }
-    }
 }
 
 #[derive(Debug, Default, Clone, PartialEq)]
@@ -35,6 +12,10 @@ pub struct EntryTags {
 }
 
 impl EntryTags {
+    pub fn new(inner: Vec<EntryTag>) -> Self {
+        Self { inner }
+    }
+
     pub fn add(&mut self, tag: EntryTag) {
         self.inner.push(tag)
     }
@@ -77,31 +58,5 @@ impl From<Vec<EntryTag>> for EntryTags {
 impl From<EntryTags> for Vec<EntryTag> {
     fn from(value: EntryTags) -> Self {
         value.inner
-    }
-}
-
-#[cfg(feature = "vdrtools_wallet")]
-impl From<EntryTags> for HashMap<String, String> {
-    fn from(value: EntryTags) -> Self {
-        let tags: Vec<EntryTag> = value.into();
-        tags.into_iter().fold(Self::new(), |mut memo, item| {
-            let (key, value) = item.into();
-
-            memo.insert(key, value);
-            memo
-        })
-    }
-}
-
-#[cfg(feature = "vdrtools_wallet")]
-impl From<HashMap<String, String>> for EntryTags {
-    fn from(value: HashMap<String, String>) -> Self {
-        Self {
-            inner: value
-                .into_iter()
-                .map(|(key, value)| (key, value))
-                .map(From::from)
-                .collect(),
-        }
     }
 }
